+++ stroom-index-shared/src/main/java/stroom/index/shared/IndexFieldType.java
@@ -19,14 +19,15 @@ package stroom.index.shared;
 import stroom.util.shared.HasDisplayValue;
 
 public enum IndexFieldType implements HasDisplayValue {
-    FIELD("Text", false),
-    ID("Id", true),
-    INT_FIELD("Int", true),
+    ID_FIELD("Id", true),
+    BOOLEAN_FIELD("Boolean", false),
+    INTEGER_FIELD("Integer", true),
     LONG_FIELD("Long", true),
     FLOAT_FIELD("Float", true),
     DOUBLE_FIELD("Double", true),
     DATE_FIELD("Date", false),
-    NUMERIC_FIELD("Number", true);
+    TEXT_FIELD("Text", false),
+    NUMERIC_FIELD("Number", true); // Alias for LONG_FIELD.
 
     private final String displayValue;
     private final boolean numeric;